chore: factor `concat_test` out of `test_common`
